argon2

Struct Argon2

pub struct Argon2<'key> { /* private fields */ }

Argon2 context.

This is the primary type of this crateโs API, and contains the following:

Argon2 Algorithm variant to be used

Argon2 Version to be used

Default set of Params to be used

(Optional) Secret key a.k.a. โpepperโ to be used

Implementations

impl<'key> Argon2<'key>

pub fn new(algorithm: Algorithm, version: Version, params: Params) -> Self

Create a new Argon2 context.

pub fn new_with_secret( secret: &'key [u8], algorithm: Algorithm, version: Version, params: Params, ) -> Result<Self>

Create a new Argon2 context.

pub fn hash_password_into( &self, pwd: &[u8], salt: &[u8], out: &mut [u8], ) -> Result<()>

Hash a password and associated parameters into the provided output buffer.

pub fn hash_password_into_with_memory( &self, pwd: &[u8], salt: &[u8], out: &mut [u8], memory_blocks: impl AsMut<[Block]>, ) -> Result<()>

Hash a password and associated parameters into the provided output buffer.

This method takes an explicit memory_blocks parameter which allows the caller to provide the backing storage for the algorithmโs state:

Users with the alloc feature enabled can use Argon2::hash_password_into to have it allocated for them.

no_std users on โheaplessโ targets can use an array of the Block type to stack allocate this buffer.

pub fn fill_memory( &self, pwd: &[u8], salt: &[u8], memory_blocks: impl AsMut<[Block]>, ) -> Result<()>

Use a password and associated parameters only to fill the given memory blocks.

This method omits the calculation of a hash and can be used when only the filled memory is required. It is not necessary to call this method before calling any of the hashing functions.

pub const fn params(&self) -> &Params

Get default configured Params.