argon2

Trait PasswordVerifier

pub trait PasswordVerifier<H>
where
    H: ?Sized,
{
    // Required method
    fn verify_password(&self, password: &[u8], hash: &H) -> Result<(), Error>;
}

Trait for password verification.

Generic around a password hash to be returned (typically phc::PasswordHash).

Automatically impl’d for type that impl PasswordHasher with phc::PasswordHash as H.

This trait is object safe and can be used to implement abstractions over multiple password hashing algorithms.

Required Methods

fn verify_password(&self, password: &[u8], hash: &H) -> Result<(), Error>

Compute this password hashing function against the provided password using the parameters from the provided password hash and see if the computed output matches.

Errors

Returns Error::Algorithm if the algorithm requested by the hash H is unsupported

Returns Error::PasswordInvalid if the computed hash for the supplied password does not match the expected hash

May return other algorithm-specific errors