Struct argon2::ParamsBuilder

pub struct ParamsBuilder { /* private fields */ }

Builder for Argon2 Params.

Implementations

impl ParamsBuilder

pub const fn new() -> Self

Create a new builder with the default parameters.

pub fn m_cost(&mut self, m_cost: u32) -> &mut Self

Set memory size, expressed in kibibytes, between 1 and (2^32)-1.

pub fn t_cost(&mut self, t_cost: u32) -> &mut Self

Set number of iterations, between 1 and (2^32)-1.

pub fn p_cost(&mut self, p_cost: u32) -> &mut Self

Set degree of parallelism, between 1 and 255.

pub fn keyid(&mut self, keyid: KeyId) -> &mut Self

Set key identifier.

pub fn data(&mut self, data: AssociatedData) -> &mut Self

Set associated data.

pub fn output_len(&mut self, len: usize) -> &mut Self

Set length of the output (in bytes).

pub const fn build(&self) -> Result<Params>

Get the finished Params.

This performs validations to ensure that the given parameters are valid and compatible with each other, and will return an error if they are not.

pub fn context( &self, algorithm: Algorithm, version: Version ) -> Result<Argon2<'_>>

Create a new Argon2 context using the provided algorithm/version.

pub const DEFAULT: ParamsBuilder = _

Default parameters (recommended).