Trait argon2::PasswordHasher

pub trait PasswordHasher {
    type Params: Clone + Debug + Default + for<'a> TryFrom<&'a PasswordHash<'a>> + TryInto<ParamsString>;
    fn hash_password_customized<'a>(
        &self, 
        password: &[u8], 
        algorithm: Option<Ident<'a>>, 
        version: Option<u32>, 
        params: Self::Params, 
        salt: impl Into<Salt<'a>>
    ) -> Result<PasswordHash<'a>, Error>;

    fn hash_password<S>(
        &self, 
        password: &[u8], 
        salt: &'a S
    ) -> Result<PasswordHash<'a>, Error>
    where
        S: AsRef<str> + ?Sized,
    { ... }
}

Trait for password hashing functions.

Associated Types

type Params: Clone + Debug + Default + for<'a> TryFrom<&'a PasswordHash<'a>> + TryInto<ParamsString>

Algorithm-specific parameters.

Required methods

fn hash_password_customized<'a>(
&self,
password: &[u8],
algorithm: Option<Ident<'a>>,
version: Option<u32>,
params: Self::Params,
salt: impl Into<Salt<'a>>
) -> Result<PasswordHash<'a>, Error>

Compute a PasswordHash from the provided password using an explicit set of customized algorithm parameters as opposed to the defaults.

When in doubt, use PasswordHasher::hash_password instead.

Provided methods

fn hash_password<S>(
&self,
password: &[u8],
salt: &'a S
) -> Result<PasswordHash<'a>, Error> where
S: AsRef<str> + ?Sized,

Simple API for computing a PasswordHash from a password and salt value.

Uses the default recommended parameters for a given algorithm.