Trait argon2::PasswordVerifier

pub trait PasswordVerifier {
    // Required method
    fn verify_password(
        &self,
        password: &[u8],
        hash: &PasswordHash<'_>
    ) -> Result<(), Error>;
}

Trait for password verification.

Automatically impl’d for any type that impls PasswordHasher.

This trait is object safe and can be used to implement abstractions over multiple password hashing algorithms. One such abstraction is provided by the PasswordHash::verify_password method.

Required Methods

fn verify_password( &self, password: &[u8], hash: &PasswordHash<'_> ) -> Result<(), Error>

Compute this password hashing function against the provided password using the parameters from the provided password hash and see if the computed output matches.